feat(br-asupersync-vks0tm): strengthen finalizer validation during region closure
This commit addresses finalizer validation gaps during region closure that could
allow finalizers to be skipped during rapid state transitions, potentially
causing resource leaks or incomplete cleanup.

**Validation improvements:**

1. **State validation during execution** - Added debug assertions to ensure
   region is in Finalizing state before and during finalizer execution
2. **TOCTOU race condition protection** - Added state re-validation after
   popping finalizers but before execution
3. **Completion tracking** - Enhanced finalizer ID tracking validation to
   catch leaked tracking state
4. **Complete close validation** - Strengthened complete_close() with state
   validation under lock to prevent race conditions
5. **Pop validation** - Added state checks in pop_finalizer() to ensure
   finalizers are only popped during valid state transitions

**Additional fixes:**
- Fixed RegionState import paths in runtime/state.rs
- Removed Copy trait from SystematicParamError (contains String field)
- Fixed type mismatches in decision_contract.rs (usize::MAX → u32::MAX)
- Added missing match arms for RfcTableInvariantViolation variant

The validation gaps were identified in asupersync-vks0tm and could potentially
allow finalizer execution to be bypassed during rapid region state transitions.
The new assertions and checks ensure finalizers are properly executed in the
correct order during the region closure protocol.

+++ src/raptorq/decision_contract.rs
@@ -458,11 +458,11 @@ fn normalize_permille_generic<const N: usize>(scores: [u32; N], zero_total: [u16
     let mut remainders = [(0usize, 0u32); N];
     let mut assigned = 0u32;
     for (index, score) in scores.iter().copied().enumerate() {
-        let scaled = score.checked_mul(PERMILLE_SCALE).unwrap_or(usize::MAX);
+        let scaled = score.checked_mul(PERMILLE_SCALE).unwrap_or(u32::MAX);
         let base = scaled / total;
         // Cap to u16::MAX if base exceeds u16 range (pathological case)
         normalized[index] = u16::try_from(base).unwrap_or(u16::MAX);
-        assigned = assigned.checked_add(base).unwrap_or(usize::MAX);
+        assigned = assigned.checked_add(base).unwrap_or(u32::MAX);
         remainders[index] = (index, scaled % total);
     }
